new orgInfo is set to result.result

+++ src/app/organization/organization-info/organization-info.component.ts
@@ -244,12 +244,21 @@ export class OrganizationInfoComponent implements OnInit {
       .subscribe(
         (orgInfo) => {
 
-          console.log('orgInfo', orgInfo);
-          this.orgInfo = orgInfo[0];
+          if (orgInfo.length > 0) {
 
-          console.log('this.orgInfo.id', this.orgInfo.id)
+            console.log('orgInfo', orgInfo);
+            this.orgInfo = orgInfo[0];
 
-          this.setFields();
+            console.log('this.orgInfo.id', this.orgInfo.id)
+
+            this.setFields();
+
+          }
+          else {
+
+            //default values
+
+          }
 
         })
 
@@ -263,7 +272,7 @@ export class OrganizationInfoComponent implements OnInit {
         (result) => {
 
           console.log('Org Info Created', result.result);
-          this.orgInfo.id = result.result.id;
+          this.orgInfo = result.result;
 
           console.log('new this.orgInfo.id', this.orgInfo.id);
 
@@ -372,9 +381,7 @@ export class OrganizationInfoComponent implements OnInit {
 
   save() {
     console.log('save pressed')
-    //the first time is create - the second time is just an update
-
-    console.log('this.orgInfo.id', this.orgInfo.id)
+    //the first time is create - the second time is a delete and create
 
     this.editing = false;
 
@@ -398,15 +405,23 @@ export class OrganizationInfoComponent implements OnInit {
 
     console.log('body', body)
 
-    this.deleteOrganizationInfoService.deleteOrgInfobyOrgInfoID(this.orgInfo.id)
-      .subscribe(
-        (result) => {
+    if (this.orgInfo) {
+      this.deleteOrganizationInfoService.deleteOrgInfobyOrgInfoID(this.orgInfo.id)
+        .subscribe(
+          (result) => {
 
-          console.log('result', result)
+            console.log('result', result)
 
-          this.createOrganizationInfo(body);
+            this.createOrganizationInfo(body);
 
-        })
+          })
+
+    }
+    else {
+
+      this.createOrganizationInfo(body);
+
+    }
 
   }
